subscribe to orderbook for depth chart
