+++ components/trade/MobileTradeAdvancedPage.tsx
@@ -11,6 +11,7 @@ import FavoriteMarketsBar from './FavoriteMarketsBar'
 
 const MobileTradeAdvancedPage = () => {
   const [activeTab, setActiveTab] = useState('trade:book')
+  const [grouping, setGrouping] = useState(0.01)
   const [showChart, setShowChart] = useState(false)
   return (
     <div className="grid grid-cols-2 sm:grid-cols-3">
@@ -41,7 +42,7 @@ const MobileTradeAdvancedPage = () => {
           />
         </div>
         <div className={activeTab === 'trade:book' ? 'visible' : 'hidden'}>
-          <Orderbook />
+          <Orderbook grouping={grouping} setGrouping={setGrouping} />
         </div>
         <div className={activeTab === 'trade:trades' ? 'visible' : 'hidden'}>
           <RecentTrades />
